Rename library uefi files to remove uefi_
This change changes the filename of any file under the library.uefi to
remove the uefi as part of the name.
