+++ chipsec/hal/common/acpi_tables.py
@@ -32,7 +32,7 @@ from collections import namedtuple
 from uuid import UUID
 from typing import List, Optional, Tuple
 from chipsec.library.logger import logger, dump_buffer_bytes
-from chipsec.library.uefi.uefi_common import EFI_GUID_FMT, EFI_GUID_STR
+from chipsec.library.uefi.common import EFI_GUID_FMT, EFI_GUID_STR
 
 
 class ACPI_TABLE:
