+++ chipsec/library/uefi/common.py
