+++ chipsec/library/uefi/compression.py
