+++ chipsec/library/uefi/platform.py
