+++ chipsec/modules/common/secureboot/variables.py
@@ -49,7 +49,7 @@ from chipsec.hal.common.uefi import UEFI
 from chipsec.library.uefi.variables import SECURE_BOOT_VARIABLES, IS_VARIABLE_ATTRIBUTE, EFI_VAR_NAME_SecureBoot, SECURE_BOOT_KEY_VARIABLES
 from chipsec.library.uefi.variables import EFI_VARIABLE_TIME_BASED_AUTHENTICATED_WRITE_ACCESS, EFI_VARIABLE_AUTHENTICATED_WRITE_ACCESS
 from chipsec.library.uefi.variables import SECURE_BOOT_OPTIONAL_VARIABLES
-from chipsec.library.uefi.uefi_common import StatusCode
+from chipsec.library.uefi.common import StatusCode
 from typing import AnyStr, List, Optional
 
 _MODULE_NAME = 'variables'
