+++ chipsec/modules/tools/uefi/scan_image.py
@@ -61,8 +61,8 @@ from chipsec.library.returncode import ModuleResult
 
 from chipsec.hal.common.uefi import UEFI
 from chipsec.hal.intel.spi import SPI, BIOS
-from chipsec.library.uefi.uefi_fv import EFI_MODULE, EFI_SECTION
-from chipsec.library.uefi.spi_uefi import build_efi_model, search_efi_tree, EFIModuleType, UUIDEncoder
+from chipsec.library.uefi.fv import EFI_MODULE, EFI_SECTION
+from chipsec.library.uefi.spi import build_efi_model, search_efi_tree, EFIModuleType, UUIDEncoder
 from chipsec.library.file import write_file, read_file
 
 TAGS = [MTAG_BIOS]
@@ -85,7 +85,7 @@ class scan_image(BaseModule):
         return True
 
     #
-    # callbacks to uefi_search.check_match_criteria
+    # callbacks to uefi.search.check_match_criteria
     #
     def genlist_callback(self, efi_module: EFI_MODULE) -> None:
         md = {}
